+++ km3pipe/io/aanet.py
@@ -16,6 +16,7 @@ import os.path
 import numpy as np
 
 from km3pipe.core import Pump, Blob
+from km3pipe.io.hdf5 import HDF5Header
 from km3pipe.dataclasses import Table
 from km3pipe.logger import get_logger
 
@@ -110,6 +111,7 @@ class AanetPump(Pump):
         self.filename = self.require('filename')
         self.ignore_hits = bool(self.get('ignore_hits'))
         self.bare = self.get('bare', default=False)
+        self.raw_header = None
         self.header = None
         self.blobs = self.blob_generator()
         self.group_id = 0
@@ -139,21 +141,24 @@ class AanetPump(Pump):
         if self.bare:
             log.info("Skipping data conversion, only passing bare aanet data")
             for event in event_file:
-                yield Blob({'evt': event})
+                yield Blob({'evt': event, 'event_file': event_file})
         else:
             log.info("Unpacking aanet header into dictionary...")
             hdr = self._parse_header(event_file.header)
             if not hdr:
                 log.info("Empty header dict found, skipping...")
-                self.header = None
+                self.raw_header = None
             else:
                 log.info("Converting Header dict to Table...")
-                self.header = self._convert_header_dict_to_table(hdr)
+                self.raw_header = self._convert_header_dict_to_table(hdr)
+                log.info("Creating HDF5Header")
+                self.header = HDF5Header.from_table(self.raw_header)
             for event in event_file:
                 log.debug('Reading event...')
                 blob = self._read_event(event, filename)
                 log.debug('Reading header...')
-                blob["RawHeader"] = self.header
+                blob["RawHeader"] = self.raw_header
+                blob["Header"] = self.header
                 self.group_id += 1
                 yield blob
         del event_file
@@ -196,6 +201,7 @@ class AanetPump(Pump):
         out = defaultdict(list)
         # iterating empty ROOT vector causes segfaults!
         if len(tracks) == 0:
+            self.log.debug("Found empty tracks, skipping...")
             return out
         for i, trk in enumerate(tracks):
             self.log.debug('Reading Track #{}...'.format(i))
@@ -312,6 +318,10 @@ class AanetPump(Pump):
 
     def _parse_mctracks(self, mctracks):
         out = defaultdict(list)
+        # iterating empty ROOT vector causes segfaults!
+        if len(mctracks) == 0:
+            self.log.debug("Found empty mctracks, skipping...")
+            return out
         for trk in mctracks:
             out['dir_x'].append(trk.dir.x)
             out['dir_y'].append(trk.dir.y)
@@ -332,6 +342,10 @@ class AanetPump(Pump):
 
     def _parse_mchits(self, mchits):
         out = defaultdict(list)
+        # iterating empty ROOT vector causes segfaults!
+        if len(mchits) == 0:
+            self.log.debug("Found empty mchits, skipping...")
+            return out
         for hit in mchits:
             out['a'].append(hit.a)
             out['origin'].append(hit.origin)
@@ -342,6 +356,10 @@ class AanetPump(Pump):
 
     def _parse_hits(self, hits):
         out = defaultdict(list)
+        # iterating empty ROOT vector causes segfaults!
+        if len(hits) == 0:
+            self.log.debug("Found empty hits, skipping...")
+            return out
         for hit in hits:
             out['channel_id'].append(hit.channel_id)
             out['dom_id'].append(hit.dom_id)
@@ -357,6 +375,8 @@ class AanetPump(Pump):
         for key, taglist in header._hdr_dict():
             tags[key] = [k for k in taglist]
         out = {}
+        if len(header) == 0:
+            return out
         for i, (key, entries) in enumerate(header):
             out[key] = {}
             for j, elem in enumerate(entries.split()):
@@ -378,6 +398,7 @@ class AanetPump(Pump):
                 out[key][elem_name] = elem
         return out
 
+    # TODO: delete this method and use the function in io/hdf5.py
     @staticmethod
     def _convert_header_dict_to_table(header_dict):
         if not header_dict: