+++ km3pipe/utils/runtable.py
@@ -11,6 +11,7 @@ Options:
     -h --help           Show this screen.
     -c                  Compact view.
     -n RUNS             Number of runs.
+    -r FROM_RUN-TO_RUN  Range of runs (example: 3100-3200).
     -s REGEX            Regular expression to filter the runsetup name/id.
     DET_ID              Detector ID (eg. D_ARCA001).
 
@@ -31,10 +32,19 @@ __status__ = "Development"
 log = kp.logger.get_logger(__name__)
 
 
-def runtable(det_id, n=5, compact=False, sep='\t', regex=None):
+def runtable(det_id, n=5, run_range=None, compact=False, sep='\t', regex=None):
     """Print the run table of the last `n` runs for given detector"""
     db = kp.db.DBManager()
     df = db.run_table(det_id)
+
+    if run_range is not None:
+        try:
+            from_run, to_run = [int(r) for r in run_range.split('-')]
+        except ValueError:
+            log.critical("Please specify a valid range (e.g. 3100-3200)!")
+            raise SystemExit
+        else:
+            df = df[(df.RUN >= from_run) & (df.RUN <= to_run)]
 
     if regex is not None:
         try:
@@ -64,4 +74,10 @@ def main():
     except TypeError:
         n = None
 
-    runtable(args['DET_ID'], n, regex=args['-s'], compact=args['-c'])
+    runtable(
+        args['DET_ID'],
+        n=n,
+        run_range=args['-r'],
+        regex=args['-s'],
+        compact=args['-c']
+    )